% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sample_priors}
\alias{sample_priors}
\title{Sample priors for MS's simulations}
\usage{
sample_priors(model_number)
}
\arguments{
\item{prior_list}{A list containing lower and upper bounds for all the priors required
by the model}
}
\value{
sampled priors from uniform distributions
}
\description{
Sample priors from a uniform distribution using the lower and upper bounds
informed by the user.
}
