% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sample_theta}
\alias{sample_theta}
\title{Sample the theta value for MS's simulations}
\usage{
sample_theta(coalescent_theta)
}
\arguments{
\item{coalescent_theta}{A boolean, if TRUE the theta is informed by the user,
if FALSE should be calculated from Ne and mutation rates.}
}
\value{
Theta values for each partition (nuclear, citoplasmatic and SSR).
}
\description{
Takes the theta values informed by the user or calculates them from Ne
and mutation rates.
}
